refactor(inventory): improve document handling and UI consistency
- Change Consecutivo type from long to int across models for consistency
- Simplify document number assignment logic
- Make document type selector and number field read-only when editing
- Improve bodega display by showing code if name is null
- Move initialization logic to OnAfterRenderAsync
- Add proper date binding and document number handling
- Enhance detail loading with proper element mapping

+++ src/Spherical.Api/Models/DocumentoTipo.cs
@@ -11,7 +11,7 @@ public partial class DocumentoTipo
 
     public string Nombre { get; set; } = null!;
 
-    public long Consecutivo { get; set; }
+    public int Consecutivo { get; set; }
 
     public string Operacion { get; set; } = null!;
