+++ src/Spherical.Client/DTO/Lineup/DocumentoTipoDto.cs
@@ -5,7 +5,7 @@ namespace Spherical.Client.DTO.Inventory
         public string Id { get; set; } = string.Empty;
         public string Empresa { get; set; } = string.Empty;
         public string Nombre { get; set; } = string.Empty;
-        public long Consecutivo { get; set; }
+        public int Consecutivo { get; set; }
         public string Operacion { get; set; } = "+"; // '+' o '-'
         public short CantidadFilas { get; set; }
         public bool EsSistema { get; set; }
